movie-booking-backend/requirements.txt
flask
flask_sqlalchemy
flask_jwt_extended
pymysql
python-dotenv
werkzeug
